requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
aiohttp>=3.8.0
python-multipart>=0.0.6
pillow>=10.0.0
httpx>=0.24.0
